requirements.txt
Flask==3.0.2
Jinja2==3.1.3
WeasyPrint==60.2
pdf2image==1.17.0
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
loguru==0.7.2 